Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float m;
-    scanf("%f",&m);
-     printf("You entered: %.4f",m);
+    double m;
+    scanf("%lf",&m);
+     printf("You entered: %.4lf",m);
     return 0;
 }
